+++ test/NCalc.Tests/LambdaTests.cs
@@ -647,8 +647,8 @@ public class LambdaTests
     public void ShouldAllowPowWithDecimals()
     {
         var e = new Expression("Pow(3.1, 2)", ExpressionOptions.DecimalAsDefault, CultureInfo.InvariantCulture);
-        var lambda = e.ToLambda<double>();
-        Assert.Equal(9.61d, lambda());
+        var lambda = e.ToLambda<decimal>();
+        Assert.Equal(9.61m, lambda());
     }
 }
 #endif
